+++ HISTORY.md
@@ -1,6 +1,7 @@
 ## 1.0.6 (2014-??-??)
 
 * Fixed an issue that made stopping of Recurrings impossible
+* Added support for registered Apps (appId / appSecret headers)
 
 ## 1.0.5 (2013-10-03)